+++ package.json
@@ -35,6 +35,7 @@
     "formik": "^2.2.9",
     "framer-motion": "^7.6.1",
     "lodash": "^4.17.21",
+    "moment": "^2.29.4",
     "react": "^18.2.0",
     "react-chartjs-2": "^4.3.1",
     "react-dom": "^18.2.0",
