+++ src/Helpers/Types/activities.ts
@@ -1,12 +1,10 @@
-interface habits {
+interface activities {
     id: number;
     title: string,
     notes: string,
     dificulty: number,
     EnDate: Date,
-    schedule: TimeRanges,
-    Time: number,
     UserId: string,
 
 }
-export default habits;
+export default activities;
