+++ src/Helpers/Types/events.ts
@@ -1,12 +1,10 @@
-interface habits {
+interface events {
     id: number;
     title: string,
     notes: string,
-    dificulty: number,
     EnDate: Date,
-    schedule: TimeRanges,
     Time: number,
     UserId: string,
 
 }
-export default habits;
+export default events;
