+++ src/Pages/Inicio/Helpers/Canva.tsx
@@ -3,7 +3,7 @@ import TileMap from './TileMap';
 
 const Canva: React.FunctionComponent = () => {
   const canvasRef = useRef<HTMLCanvasElement | null>(null);
-  const TileSize = 32;
+  const TileSize = 64;
   const tileMap = new TileMap(TileSize);
 
   useEffect(() => {
@@ -21,7 +21,7 @@ const Canva: React.FunctionComponent = () => {
     }
     tileMap.draw(canvas, context);
   // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, []);
+  }, [canvasRef]);
   return (
     <>
     <canvas ref={canvasRef}></canvas>
