--- src/Pages/Progreso/Components/Progreso.tsx
@@ -1,7 +1,0 @@
-import React from 'react';
-
-const Progreso: React.FunctionComponent = () => {
-  return <div>Progreso</div>;
-};
-
-export default Progreso;
