+++ src/hooks/useEvents.ts
@@ -1,0 +1,25 @@
+import { useSelector } from 'react-redux';
+import growApi from '../Services/Api/growApi';
+import moment from 'moment';
+
+export const useEvents: any = () => {
+
+    const { user } = useSelector((state: any) => state.auth);
+
+    const EnviarCohen = async (Answer: number): Promise<void> => {
+        try {
+            const now = moment().format("YYYY-MM-DD");
+            const resp = await growApi.post('/cohen/create', { user, Answer, now });
+            console.log(resp)
+        } catch (error) {
+            console.log({ error });
+        }
+    };
+
+
+    return {
+
+        //* Métodos
+        EnviarCohen,
+    }
+}
